when in nginx, use fcgi PROPERLY
collection.fcgi doesn't bind to a unix socket or TCP port,
which is what nginx wants.

So we were using fcgiwrap, which DOES bind, but
treats the clients as CGI, not FCGI (i.e. very inefficient).

But that was incredibly dumb, so
instead we use systemd (instead of spawn-fcgi)
to do the TINY TINY bit of the job that turns stdio into a unix
socket.

See also:

 • https://forum.nginx.org/read.php?11,4849,163020
 • https://nileshgr.com/2016/07/09/systemd-fastcgi-multiple-processes